move test code to example
